refactor(value): split Value serde + From/Index into submodules (#106)
value.rs 2853 -> 2304 lines; Value serde plumbing and From/Index moved
to value/{serde_impl,convert}.rs. Pure mechanical move, no API change
(semver-checks clean). Full suite + Miri green.
